# Conflicts:
#	src/main/java/org/choongang/global/configs/DBConn.java
#	src/main/resources/org/choongang/global/configs/mybatis-config.xml

+++ src/main/java/org/choongang/pay/entities/Pay.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @Builder
 public class Pay {
     private int studentNo;
-    private long payment;
+    private int payment;
 
 }
